+++ src/inputParser.h
@@ -23,6 +23,7 @@ typedef struct {
 
 command_t getCommand(char* line);
 
-void executeCommand(command_t command, gamma_t** g, int line);
+void executeCommand(command_t command, gamma_t** g, unsigned long line);
+
 
 #endif //GAMMA_INPUTPARSER_H
